Sync with Beta 5 of my dotfiles
THIS IS A MAJOR UPDATE that includes all sorts of important changes.
Below is the full commit message to provide you with the necessary
information.

"Prot's Dots For Debian" (PDFD) has also been updated to reflect these
changes: https://protesilaos.com/pdfd

* * *

tag v1.0.0-pdfd_beta5
Tagger: Protesilaos Stavrou <info@protesilaos.com>
Date:   Tue Jun 4 11:34:26 2019 +0300

Beta 5 of "Prot's Dots For Debian" (PDFD)

This version contains some major adjustments compared to beta version 4.
The most noticeable one is the switch from `mate-terminal` to `xterm` as
the default terminal emulator.

Another important change is the modularisation of the SXHKD configs.
They are now splits into two files: one for actions that pertain to the
Window Manager and another for WM-independent commands.  Several key
chords have been reviewed as well.

Other issues include code fixes as well as the use of bold fonts for the
focused desktop on the panel as well as the focused window in the tmux
status line.  In general these two status areas try to have the same
"visual language" to keep things consistent.

Tempus Totus has been reviewed.  It now has improved contrast between
the accept values, while keeping its look and feel.  Overall, this is my
favourite among the light themes in the Tempus collection.

Last but not least, the `bin` directory now includes two new utilities
that may or may not be useful to other users: (i) `tmr` is a simple
timer that accepts an argument that represents a valid unit of time such
as 10m for 10 minutes or 60s for 60 seconds, and (ii) `clr` which prints
the contrast ratio between two colours defined in hexadecimal RGB
notation.

Below is the full list of commits since beta 4:

    ~/dotfiles $ git log v1.0.0-pdfd_beta4..HEAD --oneline
    0538c1d Remove changing cursor shape from Vim
    17145c4 Review SXHKD for BSPWM commands
    2f76dd2 Tweak xrandr laptop adjustments
    6f2599a Fix regression in melonpanel_height pattern
    d3c5baf BREAKTHROUGH with Xterm fonts. Revert ed0c335
    adeea8f Tweak code for custom UI font selection
    0509867 Minor tweaks to Xresources
    c41e974 INTERFACE REFINEMENTS to bspwm configs
    6cac05f Update vim spell list
    4914b63 Minor update to tmux config
    dc46517 Improve melonpanel "focused" desktop styles
    171ca0f Update Tempus Totus palette
    79af8d0 Update compton config
    b81b895 Minor tweak to own_script_notify_send_mpc_status
    ab1237c Fix issue in own_script_update_environment_theme
    1708ee2 Fix regression own_script_update_environment_theme
    9559ad3 Minor tweaks: own_script_update_environment_theme
    edc0b63 Minor update to CLR script
    dd80755 NEW SCRIPT: check contrast between two colours
    0fe6c5c Feature: TMR now plays an alarm sound!
    f430518 Fix regression in tmr
    4828fe7 Fix some code issue with TMR
    275ff1b Working prototype of TMR
    613e202 Important aesthetic changes to tmux
    fa29e03 General refinements to Xresources
    ed0c335 MAJOR CHANGE: Default mono font is Fira Code
    43e7140 Minor tweaks to dunst config
    41f6f71 FIX newsboat duplicate themes
    611e8ff WORKAROUND for f0e2b4b
    f0e2b4b BREAKING CHANGE: Disable dunst shortcuts
    9f679b6 Add a couple of MPV macros to newsboat
    7727861 BREAKING CHANGE: Replace NeoMutt with Mutt
    84a5f09 Minor tweaks to XTerm resources
    d6139e6 Minor update to Xresources
    e8165ec Add Xresources-reload in SXHKD
    2826848 Update Vim colour schemes
    052d817 Enable blinking cursor shapes
    2f22fc4 Tweak XTerm cursor settings
    097ac07 Minor update to XTerm and Vim colours
    0ec7afd Build Tempus themes for XTerm and Vim
    5bbe693 Merge branch 'xtermSwitch'
    b3b5257 Minor corrections. Remove mate-terminal notes.
    6a2afaf Merge branch 'modularSXHKD'
    c1d5d1c Remove temp file
    0f61877 BREAKING CHANGES: Prepare switch to XTerm
    230e8a0 Remove temp file
    2825bab Update SXHKD modules
    098c660 Tweak bspwm external rules
    7549f63 MAJOR CHANGES: Review sxhkd and make it modular